requirements.txt
Django==1.8.3
djangorestframework==3.2.4
django-mandrill==0.1.0
mandrill==1.0.57
plivo==0.11.0
requests==2.8.1